added alternate path for metaaggreg multiple screens, but single screens can be through avg split bin lfc3d
